requirements.txt
streamlit>=1.30
pytest>=7.0
google-generativeai